+++ native/golang/cmds/ds.go
@@ -39,6 +39,9 @@ func HandleReadDatasetRequest(conn *utils.StdioConn, params []byte) (result any,
 		request.Encoding = fmt.Sprintf("IBM-%d", utils.DefaultEncoding)
 	}
 	args := []string{"data-set", "view", request.Dsname, "--encoding", request.Encoding, "--return-etag"}
+	if len(request.LocalEncoding) > 0 {
+		args = append(args, "--local-encoding", request.LocalEncoding)
+	}
 	if len(request.Volume) > 0 {
 		args = append(args, "--volser", request.Volume)
 	}
@@ -134,6 +137,9 @@ func HandleWriteDatasetRequest(conn *utils.StdioConn, params []byte) (result any
 		request.Encoding = fmt.Sprintf("IBM-%d", utils.DefaultEncoding)
 	}
 	args := []string{"data-set", "write", request.Dsname, "--encoding", request.Encoding, "--etag-only"}
+	if len(request.LocalEncoding) > 0 {
+		args = append(args, "--local-encoding", request.LocalEncoding)
+	}
 	if len(request.Etag) > 0 {
 		args = append(args, "--etag", request.Etag)
 	}