+++ native/golang/types/ds/requests.go
@@ -40,6 +40,8 @@ type ReadDatasetRequest struct {
 	Command               string `json:"command" tstype:"\"readDataset\""`
 	// Desired encoding for the dataset (optional)
 	Encoding string `json:"encoding,omitempty"`
+	// Source encoding of the dataset content (optional, defaults to UTF-8)
+	LocalEncoding string `json:"localEncoding,omitempty"`
 	// Volume serial for the data set (optional)
 	Volume string `json:"volume,omitempty"`
 	// Dataset name
@@ -53,6 +55,8 @@ type WriteDatasetRequest struct {
 	Command               string `json:"command" tstype:"\"writeDataset\""`
 	// Desired encoding for the dataset (optional)
 	Encoding string `json:"encoding,omitempty"`
+	// Source encoding of the dataset content (optional, defaults to UTF-8)
+	LocalEncoding string `json:"localEncoding,omitempty"`
 	// Last e-tag for the data set (optional, omit to overwrite)
 	Etag string `json:"etag,omitempty"`
 	// Volume serial for the data set (optional)